Refactor indexing specs.
