+++ modules/indexing/src/test/scala/ch/epfl/bluebrain/nexus/kg/indexing/schemas/SchemaIndexerSpec.scala
@@ -1,7 +1,6 @@
 package ch.epfl.bluebrain.nexus.kg.indexing.schemas
 
 import java.time.Clock
-import java.util.regex.Pattern
 
 import akka.actor.ActorSystem
 import akka.http.scaladsl.model.Uri
@@ -13,30 +12,26 @@ import cats.syntax.show._
 import ch.epfl.bluebrain.nexus.commons.http.HttpClient
 import ch.epfl.bluebrain.nexus.commons.http.HttpClient._
 import ch.epfl.bluebrain.nexus.commons.iam.acls.Meta
-import ch.epfl.bluebrain.nexus.commons.iam.identity.Caller.AnonymousCaller
 import ch.epfl.bluebrain.nexus.commons.iam.identity.Identity.Anonymous
 import ch.epfl.bluebrain.nexus.commons.sparql.client.SparqlCirceSupport._
 import ch.epfl.bluebrain.nexus.commons.sparql.client.SparqlClient
 import ch.epfl.bluebrain.nexus.commons.test._
 import ch.epfl.bluebrain.nexus.commons.types.Version
-import ch.epfl.bluebrain.nexus.kg.core.contexts.{ContextId, Contexts}
-import ch.epfl.bluebrain.nexus.kg.core.domains.{DomainId, Domains}
-import ch.epfl.bluebrain.nexus.kg.core.organizations.{OrgId, Organizations}
+import ch.epfl.bluebrain.nexus.kg.core.domains.DomainId
+import ch.epfl.bluebrain.nexus.kg.core.organizations.OrgId
 import ch.epfl.bluebrain.nexus.kg.core.schemas.SchemaEvent._
 import ch.epfl.bluebrain.nexus.kg.core.schemas.{SchemaId, SchemaName}
 import ch.epfl.bluebrain.nexus.kg.indexing.IndexingVocab.PrefixMapping._
 import ch.epfl.bluebrain.nexus.kg.indexing.Qualifier._
 import ch.epfl.bluebrain.nexus.kg.indexing.query.SearchVocab.SelectTerms._
 import ch.epfl.bluebrain.nexus.kg.indexing.{ConfiguredQualifier, IndexerFixture, Qualifier}
-import ch.epfl.bluebrain.nexus.sourcing.mem.MemoryAggregate
-import ch.epfl.bluebrain.nexus.sourcing.mem.MemoryAggregate._
 import org.apache.jena.query.ResultSet
 import org.scalatest._
 import org.scalatest.concurrent.ScalaFutures
-import ch.epfl.bluebrain.nexus.kg.core.CallerCtx._
+
 import scala.collection.JavaConverters._
 import scala.concurrent.duration._
-import scala.concurrent.{ExecutionContextExecutor, Future}
+import scala.concurrent.{ExecutionContext, Future}
 
 @DoNotDiscover
 class SchemaIndexerSpec(blazegraphPort: Int)
@@ -58,14 +53,12 @@ class SchemaIndexerSpec(blazegraphPort: Int)
   override implicit val patienceConfig: PatienceConfig =
     PatienceConfig(5 seconds, 100 milliseconds)
 
-  private implicit val ec: ExecutionContextExecutor = system.dispatcher
-  private implicit val mt: ActorMaterializer        = ActorMaterializer()
+  private implicit val ec: ExecutionContext  = system.dispatcher
+  private implicit val mt: ActorMaterializer = ActorMaterializer()
 
   private implicit val cl: UntypedHttpClient[Future] = HttpClient.akkaHttpClient
   private implicit val rs: HttpClient[Future, ResultSet] =
     HttpClient.withAkkaUnmarshaller[ResultSet]
-  private implicit val clock  = Clock.systemUTC
-  private implicit val caller = AnonymousCaller
 
   private val base                   = s"http://$localhost/v0"
   private val blazegraphBaseUri: Uri = s"http://$localhost:$blazegraphPort/blazegraph"
@@ -112,33 +105,10 @@ class SchemaIndexerSpec(blazegraphPort: Int)
 
   "A SchemaIndexer" should {
 
-    val orgsAgg = MemoryAggregate("org")(Organizations.initial, Organizations.next, Organizations.eval).toF[Future]
-
-    val domAgg =
-      MemoryAggregate("dom")(Domains.initial, Domains.next, Domains.eval)
-        .toF[Future]
-    val ctxsAgg =
-      MemoryAggregate("contexts")(Contexts.initial, Contexts.next, Contexts.eval)
-        .toF[Future]
-    val orgs = Organizations(orgsAgg)
-
-    val doms   = Domains(domAgg, orgs)
-    val ctxs   = Contexts(ctxsAgg, doms, base.toString)
     val client = SparqlClient[Future](blazegraphBaseUri)
 
-    val indexer = SchemaIndexer(client, ctxs, settings)
-    val orgRef  = orgs.create(OrgId(genId()), genJson()).futureValue
-
-    val domRef =
-      doms.create(DomainId(orgRef.id, genId()), "domain").futureValue
-    val contextId = ContextId(domRef.id, genName(), genVersion())
-
-    val replacements = Map(Pattern.quote("{{base}}") -> base, Pattern.quote("{{context}}") -> contextId.show)
-
-    val contextJson = jsonContentOf("/contexts/minimal.json", replacements)
-
-    ctxs.create(contextId, contextJson).futureValue
-    ctxs.publish(contextId, 1L).futureValue
+    val (ctxs, replacements) = createContext(base)
+    val indexer              = SchemaIndexer(client, ctxs, settings)
 
     val id   = SchemaId(DomainId(OrgId("org"), "dom"), "name", Version(1, 0, 0))
     val meta = Meta(Anonymous, Clock.systemUTC.instant())
